Initial commit - Licenses, Accessibility settings, Save system, and more!
The first pre-alpha version of Mieu Cat Chat which focuses primarily on accessibility settings as well as a save system for those accessibility settings (and more)

The save system is ported partly from Bai Lanh (My previous project) and is currently being worked on still. There seems to be a problem in if the checksum file is corrupted which leads to a crash.

Licenses have been included for the fonts which are open source, and licenses have also been included for godot, the godot dependencies at the time of commit, and the extension godotsteam which is used.

Godotsteam and other addons however are not included in this repo however, and will need to be added manually by anyone forking this repo

Current features in this build are:

-Licenses (Currently waiting on a response from steam for any to include although we do not include any steam files in this project, artist licenses are still pending for neririe and kyuu)
-Font options (size and font selector)
-FOV
-Stretch aspect picking
-Save system
-Signal bus implementation
-And more!

Currently before we get to any gameplay, focusing on accessibility is the largest priority for us as implementing it in a well thought out and efficient manner is important to us.

Project is under AGPL 3, however individual licenses which we could find have been included or the files have been excluded in case of ambiguity or lack of explicit permission.