Created the helper static class, refactoring, improved save system
Save system is slightly more modular now, having a variable which can be changed to change all the extensions in one place.
Save system also has a variable now for the encryption key
Both changes should result in less function calls overall and a faster save system

Sanity check was made to be clearer and easier to read, and also account for checksum corruption

A generic helper static class was created which currently houses the get_font function which returns a font from a font path. This reduces redundant code in rich text label and buttons.