Y sorting and Ui
Added multiple layers for nodes to be on in 2D scenes, which allows for the foreground/Ui to always be infront, and the background to always be behind.

At the same time, Y sorting was enabled for the Middleground.

All 2D scenes should follow this convention and ATLEAST include "Middleground"