Created moderation autoload
A new module which takes care of moderation (blocks/bans/kicks) has been created, which should clean up the code slightly. One of the new functions includes is_allowed() which essentially does the task of checking if a function is in the saves blocked, as well as current banned dictionaries.

Still majorly WIP