Added cryptography mode for sending messages
Messages sent with cryptography will be saved until a key is received. This allows someone to send a message (and then the recipient confirm said message) without allowing them to receive the data. This allows for lower trust operations such as a game of rock paper scissors where you don't want people to be able to change their move on the fly.