Added lost focus timer
Every 60th of a second, the game will now check if the minigame_display is visible, if it is, and it doesn't have focus, it will continue to check every 60th of a second until 1/6th of a second has passed, then it will close the minigame. This allows us to only have the minigame exist while the player is focused on it.