Fixed reset position, workarounds for godot 4.4
We could remove some of the dubious checks, and also fixed a few of the bugs in the current version of godot.

reset_position was also added so that any scene would be able to set its own default position for mieu to teleport to when clicking reset position which should fix some things.