Fixed a bug with first frame that is added being larger than others due to initialization of player
When a player joins lobby for first time, they tend to have a significantly higher number than  future values, as such we add in a 2.0 instead to simulate a normal outlier. This has lower effects due to standard deviation  highlighting outliers and most people being on the lower frame gap end compared to the 90+ frame gap from the first lag spike.