Removed some work arounds + set Movement ID to -1 once leaving scene
When a person leaves the scene for any reason (i.e logout or moving to another scene) they will set their movement ID to -1 now which should in theory prevent any additional chances of getting the same movement ID as before. This allows us to remove the RNG from last patch which improves performance while spawning a player.

Removed a workaround that was solved in 4.4