Added extra logic for spawning players in multiplayer
When creating a new character, a variable will be passed in to indicate the players name. This logic is likely to be changed in the future as the change requires sending a new string which increases latency and packet sizes, however for now it should work. More efforts shall be done in the future to update the logic to the new IP/Steam distinction in the future.